remove social user interface
